refactor: move Value component to UI dir
