+++ app/components/Activity/InvoiceModal/InvoiceModal.js
@@ -3,8 +3,7 @@ import PropTypes from 'prop-types'
 import QRCode from 'qrcode.react'
 import copy from 'copy-to-clipboard'
 import { showNotification } from 'lib/utils/notifications'
-import Value from 'components/Value'
-import { Dropdown } from 'components/UI'
+import { Dropdown, Value } from 'components/UI'
 import { FormattedDate, FormattedTime, FormattedMessage } from 'react-intl'
 import Countdown from '../Countdown'
 import messages from './messages'
