+++ app/components/Activity/PaymentModal/PaymentModal.js
@@ -2,8 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import PaperPlane from 'components/Icon/PaperPlane'
 import Zap from 'components/Icon/Zap'
-import { Dropdown } from 'components/UI'
-import Value from 'components/Value'
+import { Dropdown, Value } from 'components/UI'
 import { FormattedDate, FormattedTime, FormattedMessage } from 'react-intl'
 import messages from './messages'
 import styles from './PaymentModal.scss'
