+++ app/components/Contacts/Network/Network.js
@@ -7,8 +7,7 @@ import FaAngleDown from 'react-icons/lib/fa/angle-down'
 import { btc, blockExplorer } from 'lib/utils'
 import Plus from 'components/Icon/Plus'
 import Search from 'components/Icon/Search'
-import { BackgroundLight, Text } from 'components/UI'
-import Value from 'components/Value'
+import { BackgroundLight, Text, Value } from 'components/UI'
 
 import { FormattedNumber, FormattedMessage, injectIntl } from 'react-intl'
 import messages from './messages'
