+++ app/components/Pay/PaySummaryLightning.js
@@ -5,8 +5,7 @@ import { FormattedNumber, FormattedMessage } from 'react-intl'
 import { satoshisToFiat } from 'lib/utils/btc'
 import { decodePayReq, getNodeAlias } from 'lib/utils/crypto'
 import BigArrowRight from 'components/Icon/BigArrowRight'
-import { Bar, Dropdown, Spinner, Text, Truncate } from 'components/UI'
-import Value from 'components/Value'
+import { Bar, Dropdown, Spinner, Text, Truncate, Value } from 'components/UI'
 import { PaySummaryRow } from '.'
 import messages from './messages'
