+++ app/components/Request/RequestSummary.js
@@ -5,8 +5,7 @@ import { FormattedMessage, FormattedRelative, injectIntl } from 'react-intl'
 import { decodePayReq } from 'lib/utils/crypto'
 import { showNotification } from 'lib/utils/notifications'
 import copy from 'copy-to-clipboard'
-import { Bar, Button, Dropdown, QRCode, Text, Truncate } from 'components/UI'
-import Value from 'components/Value'
+import { Bar, Button, Dropdown, QRCode, Text, Truncate, Value } from 'components/UI'
 import { PaySummaryRow } from '../Pay'
 import messages from './messages'
