+++ app/components/UI/index.js
@@ -32,3 +32,4 @@ export TextArea from './TextArea'
 export Titlebar from './Titlebar'
 export Toggle from './Toggle'
 export Truncate from './Truncate'
+export Value from './Value'
