--- app/components/Value/index.js
@@ -1,3 +1,0 @@
-import Value from './Value'
-
-export default Value
