+++ app/components/Wallet/Wallet.js
@@ -4,9 +4,8 @@ import FaAngleUp from 'react-icons/lib/fa/angle-up'
 import FaAngleDown from 'react-icons/lib/fa/angle-down'
 import { Box, Flex } from 'rebass'
 import { btc, blockExplorer } from 'lib/utils'
-import Value from 'components/Value'
 import Settings from 'components/Settings'
-import { Button, Dropdown, Text } from 'components/UI'
+import { Button, Dropdown, Text, Value } from 'components/UI'
 import CheckAnimated from 'components/Icon/CheckAnimated'
 import ZapLogo from 'components/Icon/ZapLogo'
 import ZapLogoBlack from 'components/Icon/ZapLogoBlack'
